requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
django-sendgrid-v5==1.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
MarkupSafe==3.0.3
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
pytz==2025.2
redis==7.0.1
reportlab==4.4.7
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
stripe==13.2.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
whitenoise==6.11.0
xlrd==2.0.2
